+++ src/components/course/services/index.js
@@ -1,7 +1,0 @@
-const { CoursesServiceV2, hooks } = require('./courses.service.v2');
-
-module.exports = (app) => {
-	const courseRoute = '/users/v2/courses';
-	app.use(courseRoute, new CoursesServiceV2());
-	app.service(courseRoute).hooks(hooks);
-};